Restructure test directory
