+++ app/src/androidTestKiwix/java/org/kiwix/kiwixmobile/di/components/TestComponent.java
@@ -1,16 +1,13 @@
 package org.kiwix.kiwixmobile.di.components;
 
-import android.support.test.espresso.core.deps.dagger.Module;
 import dagger.Component;
 import javax.inject.Singleton;
 import org.kiwix.kiwixmobile.di.modules.ApplicationModule;
-import org.kiwix.kiwixmobile.di.modules.NetworkModule;
-import org.kiwix.kiwixmobile.di.modules.TestJNIModule;
 import org.kiwix.kiwixmobile.di.modules.TestJNIModule;
 import org.kiwix.kiwixmobile.di.modules.TestNetworkModule;
-import org.kiwix.kiwixmobile.utils.NetworkTest;
+import org.kiwix.kiwixmobile.tests.NetworkTest;
 import org.kiwix.kiwixmobile.utils.TestNetworkInterceptor;
-import org.kiwix.kiwixmobile.utils.ZimTest;
+import org.kiwix.kiwixmobile.tests.ZimTest;
 
 /**
  * Created by mhutti1 on 13/04/17.
