+++ app/src/androidTestKiwix/java/org/kiwix/kiwixmobile/tests/ContentTest.java
@@ -1,4 +1,4 @@
-package org.kiwix.kiwixmobile.utils;
+package org.kiwix.kiwixmobile.tests;
 
 
 import static android.support.test.espresso.Espresso.onView;
@@ -28,6 +28,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.kiwix.kiwixmobile.R;
 import org.kiwix.kiwixmobile.testutils.TestUtils;
+import org.kiwix.kiwixmobile.utils.SplashActivity;
 
 @LargeTest
 @RunWith(AndroidJUnit4.class)
